requirements.txt
streamlit>=1.32
langchain>=0.2.0
langchain-ollama>=0.1.0
azure-cosmos>=4.6.0
python-dotenv>=1.0.0
